Fix Next.js 16 compatibility issues
- Move ServerActionError class out of 'use server' file (only async functions allowed)
- Update serverComponentsExternalPackages -> serverExternalPackages
- Remove next lint in favor of eslint (deprecated in Next.js 16)
- Remove incompatible CSS package imports from globals.css
- Delete obsolete React Router hooks and vite.config files
