+++ next.config.ts
@@ -1,11 +1,7 @@
 import type { NextConfig } from 'next'
 
 const config: NextConfig = {
-  experimental: {
-    serverComponentsExternalPackages: ['usb', 'canvas'],
-    // Enable after() API for fire-and-forget printing
-    after: true,
-  },
+  serverExternalPackages: ['usb', 'canvas'],
 }
 
 export default config
