+++ package.json
@@ -12,7 +12,7 @@
     "dev": "next dev --hostname 0.0.0.0",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
-    "lint": "next lint",
+    "lint": "eslint .",
     "postinstall": "prisma generate",
     "prisma:studio": "prisma studio",
     "start": "next start -p 3030",
