--- vite.config.ts
@@ -1,8 +1,0 @@
-import { reactRouter } from "@react-router/dev/vite";
-import { defineConfig } from "vite";
-import tsconfigPaths from "vite-tsconfig-paths";
-import tailwindcss from "@tailwindcss/vite";
-
-export default defineConfig({
-  plugins: [reactRouter(), tailwindcss(), tsconfigPaths()],
-});
